pyproject.toml
[tool.poetry]
name = "Resilience Django Challenge"
version = "0.1.0"
description = "Sample subscription app for Resilience backend engineer applicants"
authors = ["Russ Snyder <russ@arceo.ai>"]

[tool.poetry.dependencies]
python = "^3.7"

Django = "^3.2.4"
djangorestframework = "^3.12.4"

[tool.poetry.dev-dependencies]
black = "21.5b2"
isort = "^5.8.0"

[tool.black]
line-length = 88
target-version = ['py37']
include = '\.pyi?$'

[tool.isort]
profile = "black"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
